fix bugs in native cache reader; getLong returns an int; modify makefile to give more warnings

+++ pixels-cache/src/main/cpp/io_pixelsdb_pixels_cache_CacheIndexSerializer.c
@@ -20,7 +20,7 @@ int getInt(const MemoryMappedFile mmap_f, long pos)
   return *addr;
 }
 
-int getLong(const MemoryMappedFile mmap_f, long pos)
+long getLong(const MemoryMappedFile mmap_f, long pos)
 {
   long *addr = (long *)(mmap_f.addr + pos);
   return *addr;
